Safari-611.1.10/LayoutTests/accessibility/mac/header-expected.txt

PASS accessibilityController.focusedElement.role is "AXRole: AXGroup"
PASS accessibilityController.focusedElement.roleDescription is "AXRoleDescription: banner"
PASS accessibilityController.focusedElement.role is "AXRole: AXGroup"
PASS accessibilityController.focusedElement.roleDescription is "AXRoleDescription: group"
PASS accessibilityController.focusedElement.role is "AXRole: AXGroup"
PASS accessibilityController.focusedElement.roleDescription is "AXRoleDescription: group"
PASS successfullyParsed is true

TEST COMPLETE
Header outside section and article elements.
Header inside an article.
Header inside a section.
